update to 1.21.2

+++ src/main/java/wtf/kity/minecraftxiv/mixin/GameRendererAccessor.java
@@ -10,7 +10,7 @@ import org.spongepowered.asm.mixin.gen.Invoker;
 @Mixin(GameRenderer.class)
 public interface GameRendererAccessor {
     @Invoker("getFov")
-    double callGetFov(Camera camera, float tickDelta, boolean changingFov);
+    float callGetFov(Camera camera, float tickDelta, boolean changingFov);
 
     @Invoker("ensureTargetInRange")
     HitResult callEnsureTargetInRange(HitResult hitResult, Vec3d cameraPos, double interactionRange);
